# src/main/java/com/vadim/tkach/gym_tracker/controller/dto/UserRegistrationCompleteDto.java
package com.vadim.tkach.gym_tracker.controller.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserRegistrationCompleteDto {
    private String token;
    private String password;
}