# src/main/resources/API.yaml
openapi: 3.0.3
info:
  title: Fitness API
  version: 1.0.0
  description: API для керування користувачами, тренуваннями, вправами та нагадуваннями.

servers:
  - url: https://localhost:8080

tags:
  - name: Users
    description: Керування користувачами
  - name: Auth
    description: Авторизація та реєстрація
  - name: Passwords
    description: Скидання і зміна пароля
  - name: Workouts
    description: Тренування користувачів
  - name: Exercises
    description: Вправи для тренувань
  - name: Workout Exercises
    description: Прив'язка вправ до тренування
  - name: Reminders
    description: Нагадування користувачів

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
      required: [id, name, email]

    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required: [name, email, password]

    UserRegistrationComplete:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]

    PasswordResetComplete:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]

paths:

  /users:
    get:
      tags: [Users]
      summary: Отримати список користувачів
      security: [ { BearerAuth: [] } ]
      responses:
        '200':
          description: Список користувачів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '204':
          description: Користувачі не знайдені

  /users/login:
    post:
      tags: [Auth]
      summary: Авторизація користувача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Успішна авторизація
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Невірні облікові дані

  /users/signup:
    post:
      tags: [Auth]
      summary: Реєстрація нового користувача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Користувач створений та email з підтвердженням надісланий

  /users/registration-confirm:
    get:
      tags: [Auth]
      summary: Підтвердити email користувача
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email підтверджено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/registration-complete:
    post:
      tags: [Auth]
      summary: Завершити реєстрацію користувача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationComplete'
      responses:
        '200':
          description: Реєстрація завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/passwords:
    put:
      tags: [Passwords]
      summary: Змінити власний пароль
      security: [ { BearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Пароль змінено
        '403':
          description: Немає прав доступу

  /users/passwords/reset:
    put:
      tags: [Passwords]
      summary: Запит на скидання пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Лист для скидання пароля надіслано

  /users/passwords/reset-verify:
    get:
      tags: [Passwords]
      summary: Перевірка токена для скидання пароля
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Токен підтверджено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/passwords/reset-complete:
    post:
      tags: [Passwords]
      summary: Завершення скидання пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetComplete'
      responses:
        '200':
          description: Пароль скинуто

  /workouts:
    get:
      tags: [Workouts]
      summary: Отримати список тренувань
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список тренувань
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workout'

    post:
      tags: [Workouts]
      summary: Створити тренування з вправами
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutInput'
      responses:
        '201':
          description: Тренування з вправами створено

  /exercises:
    get:
      tags: [Exercises]
      summary: Отримати список вправ
      responses:
        '200':
          description: Список вправ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'

    post:
      tags: [Exercises]
      summary: Додати нову вправу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseInput'
      responses:
        '201':
          description: Вправу додано

  /workout-exercises:
    post:
      tags: [Workout Exercises]
      summary: Додати вправу до тренування
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutExerciseInput'
      responses:
        '201':
          description: Вправу додано до тренування

  /reminders:
    get:
      tags: [Reminders]
      summary: Отримати список нагадувань
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список нагадувань
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'

    post:
      tags: [Reminders]
      summary: Створити нагадування
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderInput'
      responses:
        '201':
          description: Нагадування створено